+++ src/index.ts
@@ -1,17 +1,13 @@
 import Jpex from './Jpex';
-import { resolve as registerResolveFactory } from './built-ins';
 import type {
   JpexInstance,
   SetupConfig,
   NamedParameters,
-  Resolve,
   Lifecycle,
   Precedence,
 } from './types';
 
 const jpex = new Jpex() as JpexInstance;
-
-registerResolveFactory(jpex);
 
 export {
   jpex,
@@ -22,7 +18,6 @@ export type {
   JpexInstance,
   SetupConfig,
   NamedParameters,
-  Resolve,
   Precedence,
 };